extract edx_imports
